Modificar tipo de datos a float

+++ Calculadora_VictorSanchezLlada/src/main/java/examenedgit/calculadora_victorsanchezllada/Calculadora_VictorSanchezLlada.java
@@ -14,7 +14,7 @@ import java.util.Scanner;
 public class Calculadora_VictorSanchezLlada {
 
     static Scanner scanner = new Scanner(System.in);
-    static int numero1 = 0, numero2 = 0; //Variables de entrada
+    static float numero1 = 0, numero2 = 0; //Variables de entrada
     
     public static void main(String[] args) {
         
@@ -62,10 +62,10 @@ public class Calculadora_VictorSanchezLlada {
     //Método para recoger las variables de entrada
     public static void pideNumeros() {
         System.out.println("Introduce el primer número:");
-        numero1 = scanner.nextInt();
+        numero1 = scanner.nextFloat();
 
         System.out.println("Introduce el segundo número:");
-        numero2 = scanner.nextInt();
+        numero2 = scanner.nextFloat();
 
     }
 }
